requirements.txt
requests 
python-dotenv
flask